fix: handle edge case where base_type is a leaf
When base_type is a leaf, the plugin doesn't behave as expected, because leaves are skipped globally. This results in a hacky fix in which the skipping is backed out if there is a configuration with the base_type targetting leaf. This will break leaf skipping for all later rules within the CSV.